SK-3037 run contract tests for flowvault, and gate on the public API only (#397)
