+++ flowvault/src/test/java/com/skyflow/vault/data/RequestContextTests.java
@@ -1,11 +1,15 @@
 package com.skyflow.vault.data;
 
 import com.skyflow.enums.CustomHeaderKey;
+
+import java.util.Map;
 import org.junit.Assert;
 import org.junit.Test;
 
 /**
- * Batch position on the interceptor context. Without it every batch of a bulk call presents an
+ * The interceptor context: its operation, its custom headers, and the batch position.
+ *
+ * <p>Batch position matters without it every batch of a bulk call presents an
  * identical context, so a caller cannot tag them apart — no per-batch correlation id, no
  * "batch 3 of 12" logging.
  */
@@ -52,5 +56,64 @@ public class RequestContextTests {
         } catch (UnsupportedOperationException expected) {
             Assert.assertTrue(true);
         }
+    }
+
+    // ── operation and headers (moved here with the class, from common) ──────────
+    @Test
+    public void testGetOperationReturnsConstructorValue() {
+        RequestContext context = new RequestContext("INSERT");
+
+        Assert.assertEquals("INSERT", context.getOperation());
+    }
+
+    @Test
+    public void testNullOperation() {
+        RequestContext context = new RequestContext(null);
+
+        Assert.assertNull(context.getOperation());
+    }
+
+    @Test
+    public void testGetHeadersReturnsEmptyMapByDefault() {
+        RequestContext context = new RequestContext("INSERT");
+
+        Assert.assertTrue(context.getHeaders().isEmpty());
+    }
+
+    @Test
+    public void testAddHeaderIsReflectedInGetHeaders() {
+        RequestContext context = new RequestContext("INSERT");
+        context.addHeader(CustomHeaderKey.SKYFLOW_ACCOUNT_ID, "account-id-value");
+
+        Map<CustomHeaderKey, String> headers = context.getHeaders();
+
+        Assert.assertEquals(1, headers.size());
+        Assert.assertEquals("account-id-value", headers.get(CustomHeaderKey.SKYFLOW_ACCOUNT_ID));
+    }
+
+    @Test
+    public void testAddHeaderOverwritesExistingValueForSameKey() {
+        RequestContext context = new RequestContext("INSERT");
+        context.addHeader(CustomHeaderKey.SKYFLOW_ACCOUNT_ID, "first-value");
+        context.addHeader(CustomHeaderKey.SKYFLOW_ACCOUNT_ID, "second-value");
+
+        Assert.assertEquals(1, context.getHeaders().size());
+        Assert.assertEquals("second-value", context.getHeaders().get(CustomHeaderKey.SKYFLOW_ACCOUNT_ID));
+    }
+
+    @Test
+    public void testAddMultipleDistinctHeaders() {
+        RequestContext context = new RequestContext("DETOKENIZE");
+        context.addHeader(CustomHeaderKey.SKYFLOW_ACCOUNT_ID, "account-id-value");
+        context.addHeader(CustomHeaderKey.SKYFLOW_ACCOUNT_NAME, "account-name-value");
+
+        Assert.assertEquals(2, context.getHeaders().size());
+    }
+
+    @Test(expected = UnsupportedOperationException.class)
+    public void testGetHeadersReturnsUnmodifiableMap() {
+        RequestContext context = new RequestContext("INSERT");
+
+        context.getHeaders().put(CustomHeaderKey.REQUEST_ID_HEADER, "request-id-value");
     }
 }
